requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
openai>=1.0.0
llama-index
langgraph
tqdm
requests
pandas
numpy
fastapi
uvicorn[standard]
python-dotenv
supabase
openai>=1.0.0
pandas
numpy